hotfix

+++ myapp/routes/profil.js
@@ -6,9 +6,9 @@ const multer = require('multer');
 const upload = multer({ dest: 'uploads/' }); // Les fichiers seront sauvegardés dans le dossier 'uploads'
 
 
-router.get('/', function(req, res, next) {
-    result = userModel.read(req.session.user.email, function(result) {
-        result2 = piecesModel.read(req.session.user.email, function(result2) {
+router.get('/', function (req, res, next) {
+    result = userModel.read(req.session.user.email, function (result) {
+        result2 = piecesModel.read(req.session.user.email, function (result2) {
             console.log(result2);
             res.render('profil', { title: 'Page Profil Candidat', profil: result, file: result2[0] }); // Assuming you only want to pass the first file if available
         });
@@ -26,21 +26,23 @@ router.post('/update', upload.single('myFile'), function (req, res) {
         email: req.session.user.email,
     };
 
-    result = userModel.update(userProfile.email, userProfile.nom, userProfile.prenom, userProfile.tel, function (result) {
-        result2 = piecesModel.upload(userProfile.email, file, function (result2) {
-            fs.unlink(req.file.path, (err) => {
-                if (err) {
-                    console.error(err);
-                    return;
-                }
-                console.log('Fichier supprimé avec succès');
+    userModel.update(userProfile.email, userProfile.nom, userProfile.prenom, userProfile.tel, function (result) {
+        if (file) {
+            // Only try to upload a file if one was provided
+            piecesModel.upload(userProfile.email, file, function (result2) {
+                fs.unlink(req.file.path, (err) => {
+                    if (err) {
+                        console.error(err);
+                        return;
+                    }
+                    console.log('Fichier supprimé avec succès');
+                });
             });
-            res.redirect('/profil');
-        });
+        }
+        res.redirect('/profil');
     });
-
-
 });
+
